+++ rastervision/data/raster_source/geotiff_source.py
@@ -14,6 +14,7 @@ from rastervision.utils.files import download_if_needed
 log = logging.getLogger(__name__)
 wgs84 = pyproj.Proj({'init': 'epsg:4326'})
 wgs84_proj4 = '+init=epsg:4326'
+meters_per_degree = 111319.5
 
 
 def build_vrt(vrt_path, image_paths):
@@ -56,10 +57,11 @@ class GeoTiffSource(RasterioRasterSource):
 
     def _get_chip(self, window):
         no_shift = self.x_shift_meters == 0.0 and self.y_shift_meters == 0.0
-        if not no_shift:
+        yes_shift = not no_shift
+        if yes_shift:
             ymin, xmin, ymax, xmax = window.tuple_format()
-            width = xmax - xmin
-            height = ymax - ymin
+            width = window.get_width()
+            height = window.get_height()
 
             # Transform image coordinates into world coordinates
             transform = self.image_dataset.transform
@@ -71,11 +73,15 @@ class GeoTiffSource(RasterioRasterSource):
             else:
                 lon, lat = xmin2, ymin2
 
-            # Shift
+            # Shift.  This is performed by computing the shifts in
+            # meters to shifts in degrees.  Those shifts are then
+            # applied to the WGS84 coordinate.
+            #
+            # Courtesy of https://gis.stackexchange.com/questions/2951/algorithm-for-offsetting-a-latitude-longitude-by-some-amount-of-meters  # noqa
             lat_radians = math.pi * lat / 180.0
             dlon = Decimal(self.x_shift_meters) / Decimal(
-                111319.5 * math.cos(lat_radians))
-            dlat = Decimal(self.y_shift_meters) / Decimal(111319.5)
+                meters_per_degree * math.cos(lat_radians))
+            dlat = Decimal(self.y_shift_meters) / Decimal(meters_per_degree)
             lon = float(Decimal(lon) + dlon)
             lat = float(Decimal(lat) + dlat)